backend/services/heroes_service/requirements.txt
Django
djangorestframework
geopy
psycopg2
ipython
django-cors-headers